[refactor] #6 : InputField shared로 이동
