+++ src/features/find/ui/index.ts
@@ -1,3 +1,2 @@
-export * from "./InputField";
 export * from "./FindContainer";
 export * from "./GetLocationButton";
