+++ src/shared/ui/index.ts
@@ -3,3 +3,4 @@ export * from "./PlainHeader";
 export * from "./BackHeader";
 export * from "./CloseHeader";
 export * from "./LocationCard";
+export * from "./InputField";
